fixes, cleanup and a bit of documentation

+++ src/valeditor/editor/action/template/TemplateAdd.hx
@@ -27,9 +27,13 @@ class TemplateAdd extends ValEditorAction
 	
 	override public function clear():Void 
 	{
+		this.template.unregisterAction(this);
 		if (this.status == ValEditorActionStatus.UNDONE)
 		{
-			ValEditor.destroyTemplate(this.template);
+			if (this.template.canBeDestroyed())
+			{
+				ValEditor.destroyTemplate(this.template);
+			}
 		}
 		this.template = null;
 		
@@ -45,6 +49,7 @@ class TemplateAdd extends ValEditorAction
 	public function setup(template:ValEditorTemplate):Void
 	{
 		this.template = template;
+		this.template.registerAction(this);
 		this.status = ValEditorActionStatus.DONE;
 	}
